requirements.txt
pillow==6.0.0
tensorflow==1.14.0
keras==2.2.4
numpy==1.16.4
opencv-python==4.1.0.25